+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 - Dumdum client improvements:
   - Dynamically wrap message content for larger window sizes
+  - Automatically scroll message feed
 
 ## [0.2.1] - 2024-03-21